requirements.txt
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
idna==3.3
pexpect==4.8.0
ptyprocess==0.7.0
python-dotenv==0.20.0
requests==2.27.1
sqlparse==0.4.2
stripe==3.0.0
urllib3==1.26.9
